build: wire up all uses of Make A Bid to use latest
