+++ src/containers/Tables/SubdomainTable/SubdomainTable.tsx
@@ -94,7 +94,7 @@ const SubdomainTable = ({ style }: SubdomainTableProps) => {
 		<>
 			{biddingOn !== undefined && (
 				<Overlay onClose={close} open={biddingOn !== undefined}>
-					<MakeABid domain={biddingOn!} onBid={bidPlaced} />
+					<MakeABid domain={biddingOn!} onBid={bidPlaced} onClose={close} />
 				</Overlay>
 			)}
 			<GenericTable
